+++ ConvolutionalNeuralNetwork/code/matrix.hpp
@@ -176,7 +176,7 @@ public:
 		matrix& delta, 
 		matrix& momentum,
 		matrix& momentum_squared,
-		int time_step,
+		long long time_step,
 		size_t training_data_count, 
 		float learning_rate);
